prevent base64 wrapping at 76 chars and a logic fix
base64 wraps at char 76 and introduces a white space in the encoded value, assuming it's long enough. This, while rare, will result in an: unmatched "/" error when sed is used to replace the auth value. It then breaks pulling from ACR as the ACR creds are missing the auth (it remained as $auth). This will disable wrapping, thus eliminating that possibility.
Also fix a bug in logic during Profisee deployment